Дробит основной файл на модули
